# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/apps/{slug}": {
      "get": {
        "operationId": "viewAppDetailV1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewAppDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Your account is not approved for this action."
          },
          "404": {
            "description": "No app with that slug."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "apps"
        ]
      },
      "delete": {
        "operationId": "deleteAppV1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The app and its Kubernetes resources were removed."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Your account is not approved for this action."
          },
          "404": {
            "description": "No app with that slug."
          },
          "502": {
            "description": "Kubernetes teardown failed; the app was kept."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "apps"
        ]
      }
    },
...
  },
  "info": {
    "title": "Marsa API",
    "description": "Marsa PaaS HTTP API",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "marsa_session"
      }
    },
    "schemas": {
...
      "ViewAppDetailResponse": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "my-app"
          },
          "image": {
            "type": "string",
            "example": "nginx:1.27"
          },
          "url": {
            "type": "string",
            "example": "https://my-app.demo.marsa.cc"
          },
          "containerPort": {
            "type": "integer",
            "example": 80
          },
          "minReplicas": {
            "type": "integer",
            "example": 1,
            "description": "0 means the app sleeps when idle."
          },
          "maxReplicas": {
            "type": "integer",
            "example": 1
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "LOG_LEVEL": "info"
            },
            "description": "Stored environment variables; may differ from the running container until the app is redeployed."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "slug",
          "image",
          "url",
          "containerPort",
          "minReplicas",
          "maxReplicas",
          "env",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}